♻️refactor:remove justificaion-type-service
