+++ src/@core/solicitation/interfaces/index.ts
@@ -1,2 +1,1 @@
 export type { ISolicitationService } from './solicitation-service'
-export type {IJustificationTypeService} from "./justification-type-service"
