+++ src/@core/solicitation/interfaces/solicitation-service.ts
@@ -2,6 +2,7 @@ import type { ApiResponse } from '@/@core/global/responses'
 import type { SolicitationDto } from '../dtos/solicitation-dto'
 import type {
   DayOffScheduleAdjustmentSolicitationDto,
+  JustificationTypeDto,
   TimePunchLogAdjustmentSolicitationDto,
 } from '../dtos'
 
@@ -18,4 +19,13 @@ export interface ISolicitationService {
     action: 'APPROVED' | 'DENIED',
     solicitationType: 'DAY_OFF_SCHEDULE' | 'TIME_PUNCH',
   ): Promise<ApiResponse<void>>
+  listJustificationTypes(): Promise<ApiResponse<JustificationTypeDto[]>>
+  createJustificationType(
+    justificationType: JustificationTypeDto,
+  ): Promise<ApiResponse<JustificationTypeDto>>
+  updateJustificationType(
+    justificationType: JustificationTypeDto,
+    id: string,
+  ): Promise<ApiResponse<JustificationTypeDto>>
+  deleteJustificationType(id: string): Promise<ApiResponse<void>>
 }
