+++ src/api/services/index.ts
@@ -1,5 +1,4 @@
 export { AuthService } from './auth-service'
-export {JustificationTypeService} from "./justification-type-service"
 export { CollaborationService } from './collaboration-service'
 export { SolicitationService } from './solicitation-service'
 export { WorkScheduleService } from './work-schedule-service'
