+++ src/server/actions/solicitation/delete-justification-type.ts
@@ -1,12 +1,12 @@
 import { CACHE } from '@/@core/global/constants'
 import type { Action } from '@/@core/global/interfaces/rpc'
-import type { IJustificationTypeService } from '@/@core/solicitation/interfaces'
+import type { ISolicitationService } from '@/@core/solicitation/interfaces'
 
 type Request = {
   justificationTypeId: string
 }
 export const DeleteJustificationTypeAction = (
-  service: IJustificationTypeService,
+  service: ISolicitationService,
 ): Action<Request> => {
   return {
     async handle(call) {
